skip sigint test for windows

+++ test/test-signals.js
@@ -3,6 +3,7 @@
 const assert = require('assert');
 const rclnodejs = require('..');
 const childProcess = require('child_process');
+const os = require('os');
 
 function forkOnlyRemoveRclnodejsHandlers() {
   const myHandler = () => {};
@@ -46,7 +47,7 @@ if (process.env['RCLNODEJS_TEST_FORK']) {
       break;
     default:
       forkDoPublish(
-        process.argv[2] === '--non-default-context' ? new rclnodejs.Context() : undefined
+        process.argv[2] === '--non-default-context' ? new rclnodejs.Context() : undefined,
       );
   }
 } else {
@@ -62,43 +63,46 @@ if (process.env['RCLNODEJS_TEST_FORK']) {
       rclnodejs.shutdown();
     });
 
-    it('gracefully shuts downs on SIGINT when only running default context', async () => {
-      child = childProcess.fork(__filename, {
-        env: { ...process.env, RCLNODEJS_TEST_FORK: true },
-      });
-      const node = rclnodejs.createNode('test_sub');
-      node.createSubscription('std_msgs/msg/String', 'test', () => {
-        if (!child.killed) {
-          child.kill('SIGINT');
-        }
-      });
-      rclnodejs.spin(node);
-      await new Promise((res) => {
-        child.on('close', (exitCode) => {
-          assert.strictEqual(exitCode, 0);
-          res();
+    // signals is not supported in windows, see https://nodejs.org/dist/latest-v14.x/docs/api/process.html#process_signal_events
+    if (os.platform() !== 'win32') {
+      it('gracefully shuts downs on SIGINT when only running default context', async () => {
+        child = childProcess.fork(__filename, {
+          env: { ...process.env, RCLNODEJS_TEST_FORK: true },
+        });
+        const node = rclnodejs.createNode('test_sub');
+        node.createSubscription('std_msgs/msg/String', 'test', () => {
+          if (!child.killed) {
+            child.kill('SIGINT');
+          }
+        });
+        rclnodejs.spin(node);
+        await new Promise((res) => {
+          child.on('close', (exitCode) => {
+            assert.strictEqual(exitCode, 0);
+            res();
+          });
         });
       });
-    });
 
-    it('gracefully shuts downs on SIGINT when running non-default context', async () => {
-      child = childProcess.fork(__filename, ['--non-default-context'], {
-        env: { ...process.env, RCLNODEJS_TEST_FORK: true },
-      });
-      const node = rclnodejs.createNode('test_sub');
-      node.createSubscription('std_msgs/msg/String', 'test', () => {
-        if (!child.killed) {
-          child.kill('SIGINT');
-        }
-      });
-      rclnodejs.spin(node);
-      await new Promise((res) => {
-        child.on('close', (exitCode) => {
-          assert.strictEqual(exitCode, 0);
-          res();
+      it('gracefully shuts downs on SIGINT when running non-default context', async () => {
+        child = childProcess.fork(__filename, ['--non-default-context'], {
+          env: { ...process.env, RCLNODEJS_TEST_FORK: true },
+        });
+        const node = rclnodejs.createNode('test_sub');
+        node.createSubscription('std_msgs/msg/String', 'test', () => {
+          if (!child.killed) {
+            child.kill('SIGINT');
+          }
+        });
+        rclnodejs.spin(node);
+        await new Promise((res) => {
+          child.on('close', (exitCode) => {
+            assert.strictEqual(exitCode, 0);
+            res();
+          });
         });
       });
-    });
+    }
 
     it('signal handlers are removed after call removeSignalHandlers', async () => {
       // Because signal handlers is a global event and is installed on import,
